+++ packages/system/src/color/types.ts
@@ -1,0 +1,9 @@
+import { colors } from '@gympass/yoga-tokens';
+
+type ColorsValues = typeof colors | string;
+
+export type Colors = {
+  backgroundColor?: ColorsValues;
+  bg?: ColorsValues;
+  bgColor?: ColorsValues;
+};