+++ packages/system/src/overflow/index.ts
@@ -1,0 +1,2 @@
+export * from './overflow';
+export type { Overflows } from './types';