+++ packages/tokens/src/global/index.ts
@@ -1,5 +1,5 @@
 import borders from './borders';
-import breakpoints, { BREAKPOINTS_KEYS } from './breakpoints';
+import breakpoints from './breakpoints';
 import colors from './colors';
 import fonts from './fonts';
 import fontSizes from './font-sizes';
@@ -13,7 +13,6 @@ import transition from './transition';
 const tokens = {
   borders,
   breakpoints,
-  BREAKPOINTS_KEYS,
   colors,
   fonts,
   fontSizes,
@@ -29,7 +28,6 @@ export {
   tokens as default,
   borders,
   breakpoints,
-  BREAKPOINTS_KEYS,
   colors,
   fonts,
   fontSizes,
@@ -40,3 +38,5 @@ export {
   spacing,
   transition,
 };
+
+export * from './breakpoints';