+++ packages/yoga/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.114.28](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.27...@gympass/yoga@7.114.28) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.27](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.26...@gympass/yoga@7.114.27) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.26](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.25...@gympass/yoga@7.114.26) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.25](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.24...@gympass/yoga@7.114.25) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.24](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.23...@gympass/yoga@7.114.24) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.23](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.22...@gympass/yoga@7.114.23) (2024-05-28)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.22](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.21...@gympass/yoga@7.114.22) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
+## [7.114.21](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.20...@gympass/yoga@7.114.21) (2024-05-27)
+
+**Note:** Version bump only for package @gympass/yoga
+
+
+
+
+
 ## [7.114.20](https://github.com/Gympass/yoga/compare/@gympass/yoga@7.114.19...@gympass/yoga@7.114.20) (2024-05-23)
 
 **Note:** Version bump only for package @gympass/yoga